Переместил проект Дз3 из основной папки Homework в модуль task3
