+++ Homework/task3/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=task3
+spring.application.name=Homework
